* https://github.com/jessfraz/dotfiles: (64 commits)
  update for latest kernels
  bar and minus
  add pidgin
  update tests
  keypassxc
  rc.local
  cleanup
  git credentials
  update global gitignore
  remove templates
  update
  update install golang
  update config
  paths
  fix url
  fixes
  updates
  cleanup;
  name
  add yubitouch;
  ...

+++ bin/install.sh
@@ -6,7 +6,7 @@ set -e
 
 # get the user that is not root
 # TODO: makes a pretty bad assumption that there is only one other user
-USERNAME=$(find /home/* -maxdepth 0 -printf "%f" -type d)
+USERNAME=$(find /home/* -maxdepth 0 -printf "%f" -type d || echo "$USER")
 export DEBIAN_FRONTEND=noninteractive
 
 check_is_sudo() {
@@ -36,6 +36,8 @@ setup_sources() {
 	apt-get update
 	apt-get install -y \
 		apt-transport-https \
+		ca-certificates \
+		curl \
 		--no-install-recommends
 
 	add-apt-repository universe -y
@@ -46,14 +48,14 @@ setup_sources() {
 
 	cat <<-EOF >/etc/apt/sources.list
 	###### Ubuntu Main Repos
-	deb http://au.archive.ubuntu.com/ubuntu/ xenial main restricted universe multiverse 
-	deb-src http://au.archive.ubuntu.com/ubuntu/ xenial main restricted universe multiverse 
-	
+	deb http://au.archive.ubuntu.com/ubuntu/ xenial main restricted universe multiverse
+	deb-src http://au.archive.ubuntu.com/ubuntu/ xenial main restricted universe multiverse
+
 	###### Ubuntu Update Repos
-	deb http://au.archive.ubuntu.com/ubuntu/ xenial-security main restricted universe multiverse 
-	deb http://au.archive.ubuntu.com/ubuntu/ xenial-updates main restricted universe multiverse 
-	deb-src http://au.archive.ubuntu.com/ubuntu/ xenial-security main restricted universe multiverse 
-	deb-src http://au.archive.ubuntu.com/ubuntu/ xenial-updates main restricted universe multiverse 
+	deb http://au.archive.ubuntu.com/ubuntu/ xenial-security main restricted universe multiverse
+	deb http://au.archive.ubuntu.com/ubuntu/ xenial-updates main restricted universe multiverse
+	deb-src http://au.archive.ubuntu.com/ubuntu/ xenial-security main restricted universe multiverse
+	deb-src http://au.archive.ubuntu.com/ubuntu/ xenial-updates main restricted universe multiverse
 
 
 
@@ -63,9 +65,13 @@ setup_sources() {
 	# neovim
 	deb http://ppa.launchpad.net/neovim-ppa/unstable/ubuntu xenial main
 	deb-src http://ppa.launchpad.net/neovim-ppa/unstable/ubuntu xenial main
-	
-	deb http://au.archive.ubuntu.com/ubuntu xenial main 
+
+	deb http://au.archive.ubuntu.com/ubuntu xenial main
 	EOF
+
+	# yubico
+	deb http://ppa.launchpad.net/yubico/stable/ubuntu wily main
+	deb-src http://ppa.launchpad.net/yubico/stable/ubuntu wily main
 
 	# tlp: Advanced Linux Power Management
 	# http://linrunner.de/en/tlp/docs/tlp-linux-advanced-power-management.html
@@ -78,6 +84,18 @@ setup_sources() {
 	deb https://apt.dockerproject.org/repo debian-stretch experimental
 	EOF
 
+	# Add the Cloud SDK distribution URI as a package source
+	echo "deb https://packages.cloud.google.com/apt cloud-sdk-sid main" > /etc/apt/sources.list.d/google-cloud-sdk.list
+
+	# Import the Google Cloud Platform public key
+	curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -
+
+	# Add the Google Chrome distribution URI as a package source
+	echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" > /etc/apt/sources.list.d/google-chrome.list
+
+	# Import the Google Chrome public key
+	curl https://dl.google.com/linux/linux_signing_key.pub | apt-key add -
+
 	# add docker gpg key
 	apt-key adv --keyserver hkp://p80.pool.sks-keyservers.net:80 --recv-keys 58118E89F3A912897C070ADBF76221572C52609D
 
@@ -86,6 +104,12 @@ setup_sources() {
 
 	# add the neovim ppa gpg key
 	apt-key adv --keyserver hkp://p80.pool.sks-keyservers.net:80 --recv-keys 9DBB0BE9366964F134855E2255F96FCF8231B6DD
+
+	# add the yubico ppa gpg key
+	apt-key adv --keyserver hkp://p80.pool.sks-keyservers.net:80 --recv-keys 3653E21064B19D134466702E43D5C49532CBA1A9
+
+	# add the tlp apt-repo gpg key
+	apt-key adv --keyserver pool.sks-keyservers.net --recv-keys CD4E8809
 
 	# turn off translations, speed up apt-get update
 	mkdir -p /etc/apt/apt.conf.d
@@ -129,6 +153,9 @@ base() {
 		gcc \
 		git \
 		gnupg \
+		gnupg2 \
+		gnupg-agent \
+		google-cloud-sdk \
 		grep \
 		gzip \
 		hostname \
@@ -147,6 +174,7 @@ base() {
 		net-tools \
 		network-manager \
 		openvpn \
+		pinentry-curses \
 		rxvt-unicode-256color \
 		scdaemon \
 		silversearcher-ag \
@@ -171,6 +199,10 @@ base() {
 	apt-get autoremove
 	apt-get autoclean
 	apt-get clean
+
+	install_docker
+	install_scripts
+
 }
 
 # setup sudo for a user
@@ -225,14 +257,14 @@ install_docker() {
 	systemctl enable docker
 
 	# update grub with docker configs and power-saving items
-	sed -i.bak 's/GRUB_CMDLINE_LINUX=""/GRUB_CMDLINE_LINUX="cgroup_enable=memory swapaccount=1 i915.enable_psr=0 pcie_asm=force i915.i915_enable_fbc=1 i915.i915_enable_rc6=7 i915.lvds_downclock=1 apparmor=1 security=apparmor"/g' /etc/default/grub
+	sed -i.bak 's/GRUB_CMDLINE_LINUX=""/GRUB_CMDLINE_LINUX="cgroup_enable=memory swapaccount=1 pcie_aspm=force apparmor=1 security=apparmor"/g' /etc/default/grub
 	echo "Docker has been installed. If you want memory management & swap"
 	echo "run update-grub & reboot"
 }
 
 # install/update golang from source
 install_golang() {
-	export GO_VERSION=1.7.1
+	export GO_VERSION=1.7.4
 	export GO_SRC=/usr/local/go
 
 	# if we are passing the version
@@ -246,9 +278,13 @@ install_golang() {
 		sudo rm -rf "$GOPATH"
 	fi
 
-	# subshell because we `cd`
+	# subshell
 	(
-		curl -sSL "https://storage.googleapis.com/golang/go${GO_VERSION}.linux-amd64.tar.gz" | sudo tar -v -C /usr/local -xz
+	curl -sSL "https://storage.googleapis.com/golang/go${GO_VERSION}.linux-amd64.tar.gz" | sudo tar -v -C /usr/local -xz
+	local user="$USER"
+	# rebuild stdlib for faster builds
+	sudo chown -R "${user}" /usr/local/go/pkg
+	CGO_ENABLED=0 go install -a -installsuffix cgo std
 	)
 
 	# get commandline tools
@@ -263,6 +299,7 @@ install_golang() {
 	go get golang.org/x/tools/cmd/guru
 
 	go get github.com/jessfraz/apk-file
+	go get github.com/jessfraz/audit
 	go get github.com/jessfraz/bane
 	go get github.com/jessfraz/battery
 	go get github.com/jessfraz/cliaoke
@@ -280,27 +317,22 @@ install_golang() {
 
 	go get github.com/axw/gocov/gocov
 	go get github.com/brianredbeard/gpget
-	go get github.com/cloudflare/cfssl/cmd/cfssl
-	go get github.com/cloudflare/cfssl/cmd/cfssljson
 	go get github.com/crosbymichael/gistit
 	go get github.com/crosbymichael/ip-addr
-	go get github.com/cbednarski/hostess/cmd/hostess
-	go get github.com/FiloSottile/gvt
-	go get github.com/FiloSottile/vendorcheck
+	go get github.com/davecheney/httpstat
+	go get github.com/google/gops
+	go get github.com/jstemmer/gotags
 	go get github.com/nsf/gocode
 	go get github.com/rogpeppe/godef
-	go get github.com/shurcooL/git-branches
-	go get github.com/shurcooL/gostatus
 	go get github.com/shurcooL/markdownfmt
 	go get github.com/Soulou/curl-unix-socket
 
-	aliases=(cloudflare/cfssl docker/docker letsencrypt/boulder opencontainers/runc jessfraz/binctr jessfraz/contained.af)
-
+	aliases=( cloudflare/cfssl docker/docker golang/dep letsencrypt/boulder opencontainers/runc jessfraz/binctr jessfraz/contained.af )
 	for project in "${aliases[@]}"; do
 		owner=$(dirname "$project")
 		repo=$(basename "$project")
 		if [[ -d "${HOME}/${repo}" ]]; then
-			rm -rf "${HOME}/${repo}"
+			rm -rf "${HOME:?}/${repo}"
 		fi
 
 		mkdir -p "${GOPATH}/src/github.com/${owner}"
@@ -340,6 +372,17 @@ install_golang() {
 		ln -snvf "$dir" "${HOME}/${base}"
 	done
 
+	# do special things for k8s GOPATH
+	mkdir -p "${GOPATH}/src/k8s.io"
+	kubes_repos=( community kubernetes release test-infra )
+	for krepo in "${kubes_repos[@]}"; do
+		git clone "https://github.com/kubernetes/${krepo}.git" "${GOPATH}/src/k8s.io/${krepo}"
+		cd "${GOPATH}/src/k8s.io/${krepo}"
+		git remote set-url --push origin no_push
+		git remote add jessfraz "https://github.com/jessfraz/${krepo}.git"
+	done
+	)
+
 }
 
 # install graphics drivers
@@ -351,13 +394,13 @@ install_graphics() {
 		exit 1
 	fi
 
-	local pkgs="nvidia-kernel-dkms bumblebee-nvidia primus"
+	local pkgs=( nvidia-kernel-dkms bumblebee-nvidia primus )
 
 	if [[ $system == "mac" ]] || [[ $system == "dell" ]]; then
-		local pkgs="xorg xserver-xorg xserver-xorg-video-intel"
+		pkgs=( xorg xserver-xorg xserver-xorg-video-intel )
 	fi
 
-	apt-get install -y $pkgs --no-install-recommends
+	apt-get install -y "${pkgs[@]}" --no-install-recommends
 }
 
 # install custom scripts/binaries
@@ -367,6 +410,7 @@ install_scripts() {
 
 	# install speedtest
 	curl -sSL https://raw.githubusercontent.com/sivel/speedtest-cli/master/speedtest_cli.py >/usr/local/bin/speedtest
+
 	chmod +x /usr/local/bin/speedtest
 
 	# install icdiff
@@ -379,36 +423,32 @@ install_scripts() {
 	curl -sSL https://raw.githubusercontent.com/tehmaze/lolcat/master/lolcat >/usr/local/bin/lolcat
 	chmod +x /usr/local/bin/lolcat
 
-	# download syncthing binary
-	if [[ ! -f /usr/local/bin/syncthing ]]; then
-		curl -sSL https://misc.j3ss.co/binaries/syncthing >/usr/local/bin/syncthing
-		chmod +x /usr/local/bin/syncthing
-	fi
-
-	syncthing -upgrade
-
 	local scripts=(go-md2man have light)
 
 	for script in "${scripts[@]}"; do
-		curl -sSL "https://misc.j3ss.co/binaries/$script" >/usr/local/bin/$script
-		chmod +x /usr/local/bin/$script
+		curl -sSL "https://misc.j3ss.co/binaries/$script" > "/usr/local/bin/${script}"
+		chmod +x "/usr/local/bin/${script}"
 	done
 }
 
 # install syncthing
 install_syncthing() {
-	check_is_sudo
-	# Add the release PGP keys:
-	curl -s https://syncthing.net/release-key.txt | sudo apt-key add -
+	# download syncthing binary
+	if [[ ! -f /usr/local/bin/syncthing ]]; then
+		# Add the release PGP keys:
+		curl -s https://syncthing.net/release-key.txt | sudo apt-key add -
 
-	# Add the "release" channel to your APT sources:
-	echo "deb http://apt.syncthing.net/ syncthing release" | sudo tee /etc/apt/sources.list.d/syncthing.list
+		# Add the "release" channel to your APT sources:
+		echo "deb http://apt.syncthing.net/ syncthing release" | sudo tee /etc/apt/sources.list.d/syncthing.list
 
-	# Update and install syncthing:
-	apt-get update
-	apt-get install syncthing
+		# Update and install syncthing:
+		apt-get update
+		apt-get install syncthing
+	fi
 
-	curl -sSL https://raw.githubusercontent.com/jessfraz/dotfiles/master/etc/systemd/system/syncthing@.service >/etc/systemd/system/syncthing@.service
+	syncthing -upgrade
+
+	curl -sSL https://raw.githubusercontent.com/jessfraz/dotfiles/master/etc/systemd/system/syncthing@.service > /etc/systemd/system/syncthing@.service
 
 	systemctl daemon-reload
 	systemctl enable "syncthing@${USERNAME}"
@@ -434,9 +474,8 @@ install_wifi() {
 
 # install stuff for i3 window manager
 install_wmapps() {
-	local pkgs="xinit feh i3 i3lock i3status scrot neovim rofi"
-
-	apt-get install --reinstall -y $pkgs --no-install-recommends
+	local pkgs=( feh i3 i3lock i3status neovim scrot slim suckless-tools )
+	apt-get install -y "${pkgs[@]}" --no-install-recommends
 
 	# update clickpad settings
 	mkdir -p /etc/X11/xorg.conf.d/
@@ -477,8 +516,8 @@ get_dotfiles() {
 		# enable dbus for the user session
 		# systemctl --user enable dbus.socket
 
-		sudo systemctl enable i3lock@${USERNAME}
-		sudo systemctl enable suspend-sedation.service
+	sudo systemctl enable "i3lock@${USERNAME}"
+	sudo systemctl enable suspend-sedation.service
 
 		cd "$HOME"
 		mkdir -p ~/Pictures
@@ -491,61 +530,64 @@ get_dotfiles() {
 install_vim() {
 	# create subshell
 	(
-		cd "$HOME"
+	cd "$HOME"
 
-		# install .vim files
-		git clone --recursive git@github.com:jessfraz/.vim.git "$HOME/.vim"
-		ln -snf "$HOME/.vim/vimrc" "$HOME/.vimrc"
-		sudo ln -snf "$HOME/.vim" /root/.vim
-		sudo ln -snf "$HOME/.vimrc" /root/.vimrc
+	# install .vim files
+	git clone --recursive git@github.com:jessfraz/.vim.git "${HOME}/.vim"
+	ln -snf "${HOME}/.vim/vimrc" "${HOME}/.vimrc"
+	sudo ln -snf "${HOME}/.vim" /root/.vim
+	sudo ln -snf "${HOME}/.vimrc" /root/.vimrc
 
-		# alias vim dotfiles to neovim
-		mkdir -p ${XDG_CONFIG_HOME:=$HOME/.config}
-		ln -snf "/home/$USERNAME/.vim" $XDG_CONFIG_HOME/nvim
-		ln -snf "/home/$USERNAME/.vimrc" $XDG_CONFIG_HOME/nvim/init.vim
-		# do the same for root
-		sudo mkdir -p /root/.config
-		sudo ln -snf "/home/$USERNAME/.vim" /root/.config/nvim
-		sudo ln -snf "/home/$USERNAME/.vimrc" /root/.config/nvim/init.vim
+	# alias vim dotfiles to neovim
+	mkdir -p "${XDG_CONFIG_HOME:=$HOME/.config}"
+	ln -snf "${HOME}/.vim" "${XDG_CONFIG_HOME}/nvim"
+	ln -snf "${HOME}/.vimrc" "${XDG_CONFIG_HOME}/nvim/init.vim"
+	# do the same for root
+	sudo mkdir -p /root/.config
+	sudo ln -snf "${HOME}/.vim" /root/.config/nvim
+	sudo ln -snf "${HOME}/.vimrc" /root/.config/nvim/init.vim
 
-		# update alternatives to neovim
-		sudo update-alternatives --install /usr/bin/vi vi /usr/bin/nvim 60
-		sudo update-alternatives --config vi
-		sudo update-alternatives --install /usr/bin/vim vim /usr/bin/nvim 60
-		sudo update-alternatives --config vim
-		sudo update-alternatives --install /usr/bin/editor editor /usr/bin/nvim 60
-		sudo update-alternatives --config editor
+	# update alternatives to neovim
+	sudo update-alternatives --install /usr/bin/vi vi "$(which nvim)" 60
+	sudo update-alternatives --config vi
+	sudo update-alternatives --install /usr/bin/vim vim "$(which nvim)" 60
+	sudo update-alternatives --config vim
+	sudo update-alternatives --install /usr/bin/editor editor "$(which nvim)" 60
+	sudo update-alternatives --config editor
 
-		# install things needed for deoplete for vim
-		sudo apt-get update
-		sudo apt-get install -y \
-			python3-pip \
-			--no-install-recommends
-		pip3 install -U \
-			setuptools \
-			wheel \
-			neovim
+	# install things needed for deoplete for vim
+	sudo apt-get update
+
+	sudo apt-get install -y \
+		python3-pip \
+		--no-install-recommends
+
+	pip3 install -U \
+		setuptools \
+		wheel \
+		neovim
 	)
 }
 
 install_virtualbox() {
 	# check if we need to install libvpx1
 	PKG_OK=$(dpkg-query -W --showformat='${Status}\n' libvpx1 | grep "install ok installed")
-	echo Checking for libvpx1: $PKG_OK
+	echo "Checking for libvpx1: $PKG_OK"
 	if [ "" == "$PKG_OK" ]; then
 		echo "No libvpx1. Installing libvpx1."
 		jessie_sources=/etc/apt/sources.list.d/jessie.list
-		echo "deb http://httpredir.debian.org/debian jessie main contrib non-free" >$jessie_sources
+		echo "deb http://httpredir.debian.org/debian jessie main contrib non-free" > "$jessie_sources"
 
 		apt-get update
 		apt-get install -y -t jessie libvpx1 \
 			--no-install-recommends
 
 		# cleanup the file that we used to install things from jessie
-		rm $jessie_sources
+		rm "$jessie_sources"
 	fi
 
-	echo "deb http://download.virtualbox.org/virtualbox/debian vivid contrib" >>/etc/apt/sources.list.d/virtualbox.list
+	echo "deb http://download.virtualbox.org/virtualbox/debian vivid contrib" >> /etc/apt/sources.list.d/virtualbox.list
+
 	curl -sSL https://www.virtualbox.org/download/oracle_vbox.asc | apt-key add -
 
 	apt-get update
@@ -564,7 +606,7 @@ install_vagrant() {
 
 	# check if we need to install virtualbox
 	PKG_OK=$(dpkg-query -W --showformat='${Status}\n' virtualbox | grep "install ok installed")
-	echo Checking for virtualbox: $PKG_OK
+	echo "Checking for virtualbox: $PKG_OK"
 	if [ "" == "$PKG_OK" ]; then
 		echo "No virtualbox. Installing virtualbox."
 		install_virtualbox
@@ -572,12 +614,12 @@ install_vagrant() {
 
 	tmpdir=$(mktemp -d)
 	(
-		cd $tmpdir
-		curl -sSL -o vagrant.deb https://releases.hashicorp.com/vagrant/${VAGRANT_VERSION}/vagrant_${VAGRANT_VERSION}_x86_64.deb
-		dpkg -i vagrant.deb
+	cd "$tmpdir"
+	curl -sSL -o vagrant.deb "https://releases.hashicorp.com/vagrant/${VAGRANT_VERSION}/vagrant_${VAGRANT_VERSION}_x86_64.deb"
+	dpkg -i vagrant.deb
 	)
 
-	rm -rf $tmpdir
+	rm -rf "$tmpdir"
 
 	# install plugins
 	vagrant plugin install vagrant-vbguest
@@ -591,6 +633,7 @@ usage() {
 	echo "  graphics {dell,mac,lenovo}  - install graphics drivers"
 	echo "  wm                          - install window manager/desktop pkgs"
 	echo "  dotfiles                    - get dotfiles"
+	echo "  vim                         - install vim specific dotfiles"
 	echo "  golang                      - install golang and packages"
 	echo "  scripts                     - install scripts"
 	echo "  syncthing                   - install syncthing"
@@ -627,6 +670,8 @@ main() {
 		install_wmapps
 	elif [[ $cmd == "dotfiles" ]]; then
 		get_dotfiles
+	elif [[ $cmd == "vim" ]]; then
+		install_vim
 	elif [[ $cmd == "golang" ]]; then
 		install_golang "$2"
 	elif [[ $cmd == "scripts" ]]; then
